PDX-352: moving result base class to result folder
RCA: result class was in utility folder

Fix: moved result base class to result folder
