+++ src/index.ts
@@ -16,4 +16,4 @@ export { errorMessages } from './constants/errorMessages.js';
 export { sfCommandConstants } from './constants/sfCommandConstants.js';
 export { propertyFileContent } from './constants/propertyFileContent.js';
 export { commandConstants } from './constants/commandConstants.js';
-export { SFProvarResult } from './Utility/SfProvarResult.js';
+export { SFProvarResult } from './result/SfProvarResult.js';
